% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDivPlot.R
\name{betaDivPlot}
\alias{betaDivPlot}
\title{Title}
\usage{
betaDivPlot(otuTab, metaData, classToPlot, varAxes = FALSE, col = "Set3")
}
\arguments{
\item{otuTab}{otu table of your sample}

\item{metaData}{design file}

\item{classToPlot}{which column you want to plot}

\item{varAxes}{whether to plot the contributive variables (TRUE/FALSE)}

\item{col}{colour palette: including all the types of the "display.brewer.all()" in the RColorBrewer package}
}
\value{

}
\description{
Title
}
\examples{
otu_table_L2.txt <- system.file("extdata", "otu_table_L2.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
betaDivPlot(otuTab = otu_table_L2.txt, metaData = design.txt,
classToPlot = "status", col = "Set3")
}
