% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoaPlot.R
\name{pcoaPlot}
\alias{pcoaPlot}
\title{Title}
\usage{
pcoaPlot(
  distTab,
  metaData,
  distType,
  classForColor,
  classForShape = "None",
  ellipse = FALSE,
  col = "Set3",
)
}
\arguments{
\item{distTab}{distance table: bray curtis, weighted/unweighted unifrac distance table}

\item{metaData}{design file}

\item{distType}{which kind of distance you use for plot}

\item{classForColor}{color group}

\item{classForShape}{shape group}

\item{ellipse}{add 0.95 ellipse or not | TRUE or FALSE(default)}

\item{col}{colour palette: including all the types of the "display.brewer.all()" in the RColorBrewer package}
}
\value{

}
\description{
Title
}
\examples{
bray <- system.file("extdata", "bray_curtis_otu_table_css.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
pcoaPlot(bray, meta = design.txt, distType = "Bray Curtis", classForColor = "climate",classForShape = "status",
 col = "Paired")
}
