% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovaTest.R
\name{permanovaTest}
\alias{permanovaTest}
\title{Title PERMANOVA}
\usage{
permanovaTest(otuTab, metaData, var, method = "bray", permutations = 999)
}
\arguments{
\item{otuTab}{otu table of your sample}

\item{metaData}{design file}

\item{var}{variables}

\item{method}{"manhattan", "euclidean", "canberra", "clark",
"bray", "kulczynski", "jaccard", "gower", "altGower", "morisita",
"horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis",
"chisq" or "chord".}

\item{permutations}{permutation times}
}
\value{

}
\description{
Title PERMANOVA
}
\examples{
otu_table_L2.txt <- system.file("extdata", "otu_table_L2.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
permanovaTest(otu_table_L2.txt, design.txt, "status*cultivars")
}
