% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxBarPlot.R
\name{taxBarPlot}
\alias{taxBarPlot}
\title{Visualize the amplicon data}
\usage{
taxBarPlot(otuTab, metaData, classToPlot, topNum, col = "Set3")
}
\arguments{
\item{col}{}
}
\value{
otu table

design of your sample

which class you want to plot

top n taxa

colour palette, defult: Set3; Options: Palette from "display.brewer.all()" of the RColorBrewer package
}
\description{
Visualize the amplicon data
}
\examples{
otu_table_L2.txt <- system.file("extdata", "otu_table_L2.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
taxBarPlot(otuTab = otu_table_L2.txt, metaData = design.txt,
 classToPlot = "status", topNum = 10, col = "Set3")
}
