% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxBarPlot.R
\name{taxBarPlot}
\alias{taxBarPlot}
\title{Visualize the amplicon data}
\usage{
taxBarPlot(otuTab, metaData, classToPlot, topNum, col)
}
\arguments{
\item{topNum}{}
}
\value{
barplot of the microbial composition
}
\description{
Visualize the amplicon data
}
\examples{
otu_table_L2.txt <- system.file("extdata", "otu_table_L2.txt", package = "microVisu")
design.txt <- system.file("extdata", "design.txt", package = "microVisu")
taxBarPlot(otuTab = otu_table_L2.txt, metaData = design.txt,
 classToPlot = "status", topNum = 10, col = "Set3")
}
